microservice/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==1.2.2

# Core dependencies

starlette==1.0.0
pydantic==2.12.5
pydantic-core==2.41.5
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0
annotated-doc==0.0.4
anyio==4.13.0
idna==3.11
h11==0.16.0
click==8.3.1
colorama==0.4.6

# LangChain + LangGraph stack

langchain==1.2.14
langchain-core==1.2.23
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23

# HTTP / utils (used internally but pinned for safety)

httpx==0.28.1
httpcore==1.0.9
requests==2.33.1
requests-toolbelt==1.0.0
urllib3==2.6.3
certifi==2026.2.25
charset_normalizer==3.4.6

# Serialization / performance

orjson==3.11.8
ormsgpack==1.12.2
zstandard==0.25.0
xxhash==3.6.0

# Utilities

pyyaml==6.0.3
packaging==26.0
tenacity==9.1.4
jsonpatch==1.33
jsonpointer==3.1.1
uuid-utils==0.14.1


langchain-google-genai